backend/requirements.txt

# Python dependencies for backend
Django>=4.0
djangorestframework
django-cors-headers
psycopg2-binary

# Advanced NLP and monitoring
transformers
sentence-transformers
langdetect
dateparser
sentry-sdk
